+++ packages/sdk/src/constants/contract-features.ts
@@ -33,6 +33,7 @@ import {
 } from "./erc1155-features";
 import {
   FEATURE_METADATA,
+  FEATURE_OWNER,
   FEATURE_PERMISSIONS,
   FEATURE_PLATFORM_FEE,
   FEATURE_PRIMARY_SALE,
@@ -76,7 +77,8 @@ export type Feature =
   | typeof FEATURE_PRIMARY_SALE
   | typeof FEATURE_PERMISSIONS
   | typeof FEATURE_METADATA
-  | typeof FEATURE_APPURI;
+  | typeof FEATURE_APPURI
+  | typeof FEATURE_OWNER;
 
 /**
  * @internal
@@ -103,4 +105,5 @@ export const SUPPORTED_FEATURES: Record<string, Feature> = {
   [FEATURE_PERMISSIONS.name]: FEATURE_PERMISSIONS,
   [FEATURE_METADATA.name]: FEATURE_METADATA,
   [FEATURE_APPURI.name]: FEATURE_APPURI,
+  [FEATURE_OWNER.name]: FEATURE_OWNER,
 };