+++ packages/sdk/src/contracts/smart-contract.ts
@@ -2,7 +2,18 @@ import { ALL_ROLES, assertEnabled, detectContractFeature } from "../common";
 import { FEATURE_TOKEN } from "../constants/erc20-features";
 import { FEATURE_NFT } from "../constants/erc721-features";
 import { FEATURE_EDITION } from "../constants/erc1155-features";
-import { ContractEncoder, NetworkOrSignerOrProvider } from "../core";
+import {
+  FEATURE_OWNER,
+  FEATURE_PERMISSIONS,
+  FEATURE_PLATFORM_FEE,
+  FEATURE_PRIMARY_SALE,
+  FEATURE_ROYALTY,
+} from "../constants/thirdweb-features";
+import {
+  ContractEncoder,
+  ContractOwner,
+  NetworkOrSignerOrProvider,
+} from "../core";
 import { AppURI } from "../core/classes/appuri";
 import { ContractEvents } from "../core/classes/contract-events";
 import { ContractInterceptor } from "../core/classes/contract-interceptor";
@@ -26,6 +37,7 @@ import type {
   IPlatformFee,
   IPrimarySale,
   IRoyalty,
+  Ownable,
 } from "@thirdweb-dev/contracts-js";
 import { IStorage } from "@thirdweb-dev/storage";
 import { BaseContract, CallOverrides, ContractInterface } from "ethers";
@@ -45,10 +57,10 @@ import { BaseContract, CallOverrides, ContractInterface } from "ethers";
  * await contract.call("myCustomFunction", param1, param2);
  *
  * // if your contract follows the ERC721 standard, contract.nft will be present
- * const allNFTs = await contract.nft.query.all()
+ * const allNFTs = await contract.erc721.query.all()
  *
  * // if your contract extends IMintableERC721, contract.nft.mint() will be available
- * const tx = await contract.nft.mint({
+ * const tx = await contract.erc721.mint({
  *     name: "Cool NFT",
  *     image: readFileSync("some_image.png"),
  *   });
@@ -61,7 +73,6 @@ export class SmartContract<TContract extends BaseContract = BaseContract>
 {
   private contractWrapper;
   private storage;
-  private options;
 
   // utilities
   public events: ContractEvents<TContract>;
@@ -70,37 +81,62 @@ export class SmartContract<TContract extends BaseContract = BaseContract>
   public estimator: GasCostEstimator<TContract>;
   public publishedMetadata: ContractPublishedMetadata<TContract>;
   public abi: ContractInterface;
-
-  // features
   public metadata: ContractMetadata<BaseContract, any>;
-  public royalties: ContractRoyalty<IRoyalty, any> | undefined;
-  public roles: ContractRoles<IPermissions, any> | undefined;
-  public sales: ContractPrimarySale<IPrimarySale> | undefined;
-  public platformFees: ContractPlatformFee<IPlatformFee> | undefined;
 
-  private token: Erc20 | undefined;
-  private nft: Erc721 | undefined;
-  private edition: Erc1155 | undefined;
+  /**
+   * Handle royalties
+   */
+  get royalties(): ContractRoyalty<IRoyalty, any> {
+    return assertEnabled(this.detectRoyalties(), FEATURE_ROYALTY);
+  }
+
+  /**
+   * Handle permissions
+   */
+  get roles(): ContractRoles<IPermissions, any> {
+    return assertEnabled(this.detectRoles(), FEATURE_PERMISSIONS);
+  }
+
+  /**
+   * Handle primary sales
+   */
+  get sales(): ContractPrimarySale<IPrimarySale> {
+    return assertEnabled(this.detectPrimarySales(), FEATURE_PRIMARY_SALE);
+  }
+
+  /**
+   * Handle platform fees
+   */
+  get platformFees(): ContractPlatformFee<IPlatformFee> {
+    return assertEnabled(this.detectPlatformFees(), FEATURE_PLATFORM_FEE);
+  }
+
+  /**
+   * Set and get the owner of the contract
+   */
+  get owner(): ContractOwner<Ownable> {
+    return assertEnabled(this.detectOwnable(), FEATURE_OWNER);
+  }
 
   /**
    * Auto-detects ERC20 standard functions.
    */
   get erc20(): Erc20 {
-    return assertEnabled(this.token, FEATURE_TOKEN);
+    return assertEnabled(this.detectErc20(), FEATURE_TOKEN);
   }
 
   /**
    * Auto-detects ERC721 standard functions.
    */
   get erc721(): Erc721 {
-    return assertEnabled(this.nft, FEATURE_NFT);
+    return assertEnabled(this.detectErc721(), FEATURE_NFT);
   }
 
   /**
    * Auto-detects ERC1155 standard functions.
    */
   get erc1155(): Erc1155 {
-    return assertEnabled(this.edition, FEATURE_EDITION);
+    return assertEnabled(this.detectErc1155(), FEATURE_EDITION);
   }
 
   constructor(
@@ -116,7 +152,6 @@ export class SmartContract<TContract extends BaseContract = BaseContract>
       options,
     ),
   ) {
-    this.options = options;
     this.storage = storage;
     this.contractWrapper = contractWrapper;
     this.abi = abi;
@@ -135,17 +170,6 @@ export class SmartContract<TContract extends BaseContract = BaseContract>
       CustomContractSchema,
       this.storage,
     );
-
-    // feature detection
-    this.royalties = this.detectRoyalties();
-    this.roles = this.detectRoles();
-    this.sales = this.detectPrimarySales();
-    this.platformFees = this.detectPlatformFees();
-
-    this.token = this.detectErc20();
-    this.nft = this.detectErc721();
-    this.edition = this.detectErc1155();
-    this.appuri = this.detectAppURI();
   }
 
   onNetworkUpdated(network: NetworkOrSignerOrProvider): void {
@@ -254,5 +278,11 @@ export class SmartContract<TContract extends BaseContract = BaseContract>
     // ContractMetadata is stateless, it's fine to create a new one here
     // This also makes it not order dependent in the feature detection process
     return new AppURI(this.contractWrapper);
+  }
+  private detectOwnable() {
+    if (detectContractFeature<Ownable>(this.contractWrapper, "Ownable")) {
+      return new ContractOwner(this.contractWrapper);
+    }
+    return undefined;
   }
 }